+++ Documentation/NuttX.html
@@ -6859,7 +6859,7 @@ BFD_ASSERT (*plt_offset != (bfd_vma) -1);
         The initial release of support for the z16f was made available in NuttX version 0.3.7.
         A working NuttShell (NSH) configuration as added in NuttX-6.33 (although a patch is required to work around an issue with a ZDS-II 5.0.1 tool problem).
         An ESPI driver was added in NuttX-7.2.
-        Refer to the NuttX board <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f2811/z16f2800100zcog/README.txt" target="_blank">README</a> file for further information.
+        Refer to the NuttX board <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f/z16f2800100zcog/README.txt" target="_blank">README</a> file for further information.
       </p>
     </li>
   </ul>
