+++ Documentation/README.html
@@ -467,12 +467,12 @@ nuttx/
  |   |       `- esp32-core/
  |   |           `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/xtensa/esp32/esp32-core/README.txt" target="_blank"><b><i>README.txt</i></b></a>
  |   |- z16/
- |   |   `- z16f2811/
+ |   |   `- z16f/
  |   |       `- z16f2800100zcog/
- |   |           |- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f2811/z16f2800100zcog/configs/nsh/README.txt" target="_blank"><b><i>configs/nsh/README.txt</i></b></a>
- |   |           |- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f2811/z16f2800100zcog/configs/ostest/README.txt" target="_blank"><i>configs/ostest/README.txt</i></a>
- |   |           |- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f2811/z16f2800100zcog/configs/pashello/README.txt" target="_blank"><i>configs/pashello/README.txt</i></a>
- |   |           `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f2811/z16f2800100zcog/README.txt" target="_blank"><b><i>README.txt</i></b></a>
+ |   |           |- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f/z16f2800100zcog/configs/nsh/README.txt" target="_blank"><b><i>configs/nsh/README.txt</i></b></a>
+ |   |           |- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f/z16f2800100zcog/configs/ostest/README.txt" target="_blank"><i>configs/ostest/README.txt</i></a>
+ |   |           |- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f/z16f2800100zcog/configs/pashello/README.txt" target="_blank"><i>configs/pashello/README.txt</i></a>
+ |   |           `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/z16/z16f/z16f2800100zcog/README.txt" target="_blank"><b><i>README.txt</i></b></a>
  |   |- z80/
  |   |   |- ez80/
  |   |   |   |- ez80f910200kitg/
