+++ boards/z16/z16f/z16f2800100zcog/configs/nsh/README.txt
