+++ boards/z16/z16f/z16f2800100zcog/configs/nsh/nsh.zfpproj
