+++ boards/z16/z16f/z16f2800100zcog/configs/ostest/defconfig
@@ -5,44 +5,37 @@
 # You can then do "make savedefconfig" to generate a new defconfig file that includes your
 # modifications.
 #
-# CONFIG_NSH_ARGCAT is not set
-# CONFIG_NSH_CMDOPT_DF_H is not set
-# CONFIG_NSH_CMDOPT_HEXDUMP is not set
-# CONFIG_NSH_CMDPARMS is not set
-# CONFIG_NSH_DISABLE_IFCONFIG is not set
-# CONFIG_NSH_DISABLE_PS is not set
 CONFIG_ARCH_BOARD_Z16F2800100ZCOG=y
 CONFIG_ARCH_BOARD="z16f2800100zcog"
-CONFIG_ARCH_CHIP="z16f2811"
+CONFIG_ARCH_CHIP="z16f"
 CONFIG_ARCH_CHIP_Z16F=y
 CONFIG_ARCH_CHIP_Z16F2811=y
 CONFIG_ARCH_Z16=y
 CONFIG_ARCH="z16"
 CONFIG_BOARD_LOOPSPERMSEC=1250
-CONFIG_BUILTIN=y
+CONFIG_DEV_LOWCONSOLE=y
 CONFIG_ENDIAN_BIG=y
-CONFIG_SYSTEM_NSH=y
 CONFIG_HOST_WINDOWS=y
+CONFIG_IDLETHREAD_STACKSIZE=4096
 CONFIG_MAX_TASKS=16
 CONFIG_MAX_WDOGPARMS=2
 CONFIG_NFILE_DESCRIPTORS=8
 CONFIG_NFILE_STREAMS=8
-CONFIG_NSH_BUILTIN_APPS=y
-CONFIG_NSH_FILEIOSIZE=512
-CONFIG_NSH_READLINE=y
 CONFIG_PREALLOC_MQ_MSGS=4
 CONFIG_PREALLOC_TIMERS=4
-CONFIG_PREALLOC_WDOGS=16
-CONFIG_PTHREAD_STACK_DEFAULT=1536
+CONFIG_PREALLOC_WDOGS=4
+CONFIG_PTHREAD_STACK_DEFAULT=4096
 CONFIG_RAM_SIZE=65536
 CONFIG_RR_INTERVAL=200
-CONFIG_SCHED_WAITPID=y
 CONFIG_SDCLONE_DISABLE=y
-CONFIG_START_YEAR=2014
+CONFIG_START_DAY=29
+CONFIG_START_MONTH=11
+CONFIG_START_YEAR=2012
 CONFIG_TASK_NAME_SIZE=0
-CONFIG_TASK_SPAWN_DEFAULT_STACKSIZE=1536
+CONFIG_TESTING_OSTEST=y
 CONFIG_UART0_BAUD=57600
 CONFIG_UART0_SERIAL_CONSOLE=y
 CONFIG_UART1_BAUD=57600
-CONFIG_USER_ENTRYPOINT="nsh_main"
-CONFIG_WDOG_INTRESERVE=2
+CONFIG_USER_ENTRYPOINT="ostest_main"
+CONFIG_USERMAIN_STACKSIZE=4096
+CONFIG_WDOG_INTRESERVE=0
